refactor(netgrep): share the tail ceiling, the engine gate and two helpers
