+++ packages/netgrep/src/lib/concatBytes.ts
@@ -1,0 +1,16 @@
+/**
+ * Join a list of byte chunks into one buffer.
+ */
+export function concatBytes(chunks: Array<Uint8Array>): Uint8Array {
+  const total = chunks.reduce((sum, chunk) => sum + chunk.length, 0);
+  const joined = new Uint8Array(total);
+
+  let offset = 0;
+
+  for (const chunk of chunks) {
+    joined.set(chunk, offset);
+    offset += chunk.length;
+  }
+
+  return joined;
+}
